Assassin-dev/requirements.txt
asgiref==3.4.1
boto3==1.20.24
botocore==1.23.24
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
django-background-tasks==1.2.5
django-compat==1.0.15
django-rest-framework==0.1.0
djangorestframework==3.13.1
ffmpeg==1.4
idna==3.3
jmespath==0.10.0
mysqlclient==2.1.0
numpy==1.21.5
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
stripe==2.63.0
tzdata==2021.5
urllib3==1.26.7
